+++ lib/active_interaction/modules/core.rb
@@ -4,10 +4,9 @@ rescue LoadError
 end
 
 module ActiveInteraction
-  # Functionality common between {Base} and {Pipeline}.
+  # Functionality common between {Base}.
   #
   # @see Base
-  # @see Pipeline
   module Core
     # Get or set the description.
     #