+++ CHANGELOG.md
@@ -1,5 +1,7 @@
 # [Master][]
 
+- Add experimental composition implementation
+  (`ActiveInteraction::Base#compose`).
 - Remove `ActiveInteraction::Pipeline`.
 
 # [0.8.0][] (2013-11-14)